\name{geneInfo}
\alias{geneInfo}
\docType{data}
\title{
  Gene information
}
\description{
  Gene annotation (SGD, release 64) for the RNA-Seq experiment of Risso et al. (2011).

  This is a data.frame with two columns:

  1. length: gene length in base pairs;

  2. GC: proportion of G and C nucleotides in the gene region.
}
\usage{data(geneInfo)}
\source{
  A citation for the dataset may be generated using
  \code{citation("yeastGeneLevel")}. 
}
\keyword{datasets}