\name{geneLevelCounts}
\alias{geneLevelCounts}
\alias{yeastGeneLevel}
\docType{data}
\title{
  Yeast gene-level counts from: Risso et al. (2011) BMC Bioinformatics.
}
\description{
  Gene counts for an Illumina GA II sequencing experiment. Reads were aligned to the reference genome (SGD release 64, www.yeastgenome.org).
  The read count for a given gene is defined as the number of reads with 5'-end falling within the corresponding region, following SGD annotation.
}
\usage{data(geneLevelCounts)}
\source{
  A citation for the dataset may be generated using
  \code{citation("yeastGeneLevel")}. 
}
\keyword{datasets}
