\name{laneInfo}
\alias{laneInfo}
\docType{data}
\title{
  Lane information
}
\description{
  Lane annotation for the RNA-Seq experiment of Risso et al. (2011).

  This is a data.frame with four columns:

  1. libr_prep: a unique name for the colony/library preparation;

  2. growth_cond: growth condition, standard YP Glucose (YPD), Delft Glucose (Del) and YP Glycerol (Gly);

  3. flow_cell: the flow cell in which the sample has been sequenced;

  4. lib_prep_proto: library preparation protocol (see Risso et al. (2011) for details).
}
\usage{data(laneInfo)}
\source{
  A citation for the dataset may be generated using
  \code{citation("yeastGeneLevel")}. 
}
\keyword{datasets}